+++ vegetation/batch.py
@@ -1,11 +1,6 @@
 from mesa.batchrunner import batch_run
-from vegetation.patch.model import Vegetation
+from vegetation.model.vegetation import Vegetation
 from numpy import arange
-from vegetation.config.paths import (
-    LOCAL_STAC_CACHE_FSTRING,
-    SAVE_LOCAL_STAC_CACHE,
-    DEM_STAC_PATH,
-)
 
 # TODO: Implement early stopping when all the JOTR die off
 # Issue URL: https://github.com/SchmidtDSE/mesa_abm_poc/issues/18
